Fix import error
Comment out `moduleResolution` in tsconfig.json because when `module` is
set to an ES module style, `moduleResolution` defaults to `classic`,
which matches `"type": "module"` in package.json. Refer the officail
document for more details.
https://www.typescriptlang.org/tsconfig#moduleResolution

Add ".ts" and ".tsx" into the rules of "import/extensions" to never
import typescript files with explicitly spelling out the file extensions
of the typescript files to match the code style of Airbnb.