Reimplement data types in typescript
In the previous commit, `moduleResolution` typescript option is set to
the default value, `classic`, which is not compatible with the module
resolution of Node.js, so this commit rolls it back to "node" and adds
`eslint-import-resolver-typescript` the module to handle the module
resolution of typescript. Refer to https://segmentfault.com/a/1190000040261998
for more details.

Move the crawler for downloading data of the Taoyuan City (TYC) to
`src/crawlers`.

Add some ESLint rules for typescript. The options of the rules set by
airbnb-base are preserved.