Enumerate case severity
Add an enumerator for case severities.

Reduce the field checked in the method to check if the two `Case`s are
the same one. The original version and this new version are not perfect
to distinguish different cases. Maybe we should distinguish different
cases by the structure of the dataset.

#3