Replace case ID and party ID with UUID
Multiple cases may happen at the same time and location, so we cannot
use the combination of date and location to identify the case. The
easiest way to identify the case is to use UUID.

Also add unique UUID as the party ID to identify the party in Google
Data Studio.

#3