Add table schemas using on BigQuery
`Vehicle` is a required parameter in `Party`. Logically, the BigQuery
field corresponding to `Vehicle` should also be set as a required field.
However, when exporting the cases into CSV, there is one possible value
of `Vehicle` that is converted to an empty string. Empty string
represents NULL in BigQuery. So we set `Vehicle` as NULLABLE.

Issue: #8