Read CSV stream in `class BaseParser`
Every parser will use the same code to read the CSV stream, so it is
good to move the code reading the CSV stream to the base class.

Turn off `no-underscore-dangle` ESLint rule. Airbnb enables this rule
because javascript doesn't have a concept of private and protected.
However, in Typescript, that concept exists. "_" is useful to make the
name of a private or protected variable different from its setter and
getter.

Issue: #5